fix: move logger to pkg
