+++ internal/datautils/map.go
@@ -8,8 +8,8 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/steebchen/prisma-client-go/runtime/types"
 
-	"github.com/hatchet-dev/hatchet/internal/logger"
 	"github.com/hatchet-dev/hatchet/pkg/errors"
+	"github.com/hatchet-dev/hatchet/pkg/logger"
 	"github.com/hatchet-dev/hatchet/pkg/validator"
 )
