+++ pkg/worker/worker.go
@@ -10,11 +10,11 @@ import (
 
 	"github.com/rs/zerolog"
 
-	"github.com/hatchet-dev/hatchet/internal/logger"
 	"github.com/hatchet-dev/hatchet/pkg/client"
 	"github.com/hatchet-dev/hatchet/pkg/client/types"
 	"github.com/hatchet-dev/hatchet/pkg/errors"
 	"github.com/hatchet-dev/hatchet/pkg/integrations"
+	"github.com/hatchet-dev/hatchet/pkg/logger"
 )
 
 type actionFunc func(args ...any) []any
